+++ packages/cli/src/list/data-set-members/DataSetMembers.handler.ts
@@ -17,6 +17,7 @@ export default class ListDataSetMembersHandler extends SshBaseHandler {
     public async processWithClient(params: IHandlerParameters, client: ZSshClient): Promise<ds.ListDsMembersResponse> {
         const response = await client.ds.listDsMembers({
             dsname: params.arguments.dsname,
+            attributes: params.arguments.attributes,
             pattern: params.arguments.pattern,
         });
         params.response.data.setMessage(
@@ -25,9 +26,12 @@ export default class ListDataSetMembersHandler extends SshBaseHandler {
             params.arguments.dsname,
         );
         params.response.format.output({
-            output: response.items,
+            output: response.items.map((item) => ({
+                ...item,
+            })),
             format: "table",
-            fields: ["name"],
+            fields: ["name", "vers", "mod", "c4date", "m4date", "mtime", "cnorc", "inorc", "mnorc", "user", "sclm"],
+            header: true,
         });
         return response;
     }